# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Naytia 像素小站 API",
  "description": "账号 / 论坛 / 工具箱收藏。契约见 docs/plans/03。",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/admin/accounts": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Accounts",
    "description": "账号列表，可选按用户名模糊搜索（`q`，大小写不敏感的子串匹配）。\n\n搜索放在服务端而不是前端过滤：这一页是**管理**界面，账号数会随时间增长，\n而「搜不到」在管理场景下是错误结论（以为这个人不存在）。用 SQL 过滤，\n前端拿到的永远是与搜索词匹配的那一批，不依赖已加载的那一页。\n\n不传 `q` 时仍然按 id 升序返回全部——注册最早的账号排在最前，\n也就是超管自己（引导规则，ADR-002）在最上面。\n\n`max_length` 给得比用户名上限宽：搜索词比任何用户名都长时结果自然为空，\n那是个正常结果，不该回一个 422 让使用者去猜自己做错了什么。",
    "operationId": "list_accounts_api_admin_accounts_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 64
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/AccountSummary"
         },
         "title": "Response List Accounts Api Admin Accounts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/accounts/{account_id}/admin": {
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Toggle Admin",
    "description": "授予 / 收回管理员（admin ↔ member）。\n\n单独开一个接口而不是让前端去调 `role` 那个：`role` 接口按契约收任何角色，\n一个只该「开关管理员」的界面不该顺手拥有把别人提成超管的能力。\n超管档位只由引导规则与邀请码产生（ADR-002）。",
    "operationId": "toggle_admin_api_admin_accounts__account_id__admin_put",
    "parameters": [
     {
      "name": "account_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Account Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleToggle"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/admin/accounts/{account_id}/role": {
   "put": {
    "tags": [
     "admin"
    ],
    "summary": "Update Role",
    "operationId": "update_role_api_admin_accounts__account_id__role_put",
    "parameters": [
     {
      "name": "account_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Account Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AccountSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AccountSummary": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "username",
     "role",
     "created_at"
    ],
    "title": "AccountSummary"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RoleToggle": {
    "properties": {
     "enabled": {
      "type": "boolean",
      "title": "Enabled",
      "description": "true = 授予 admin；false = 退回 member"
     }
    },
    "type": "object",
    "required": [
     "enabled"
    ],
    "title": "RoleToggle",
    "description": "「设为管理员 / 取消管理员」的语义化入参（/api/admin/accounts/{id}/admin）。\n\n为什么不复用 RoleUpdate：那个接口收的是**任意角色**，包括 superadmin——\n一个「给不给管理员」的开关，不应该具备把别人提成超管的能力。\n这里的 `enabled` 只映射到 admin 与 member 两档，超管只能由引导规则或邀请码产生。"
   },
   "RoleUpdate": {
    "properties": {
     "role": {
      "type": "string",
      "pattern": "^(member|admin|superadmin)$",
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "role"
    ],
    "title": "RoleUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}